nouveau: notifier timeout is a float now

+++ src/gallium/drivers/nouveau/nouveau_winsys.h
@@ -50,7 +50,7 @@ struct nouveau_winsys {
 	uint32_t  (*notifier_status)(struct nouveau_notifier *, int id);
 	uint32_t  (*notifier_retval)(struct nouveau_notifier *, int id);
 	int       (*notifier_wait)(struct nouveau_notifier *, int id,
-				   int status, int timeout);
+				   int status, float timeout);
 
 	int (*surface_copy)(struct nouveau_winsys *, struct pipe_surface *,
 			    unsigned, unsigned, struct pipe_surface *,
